Extracted out several views / extensions to make files smaller / easier to read
