+++ Shared/Views/MobSessionView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 struct MobSessionView: View {
     @AppStorage("firstTimeUser") var showingInfoSheet = true
     @EnvironmentObject var vm: MobSessionManager
-    @State private var editMode = EditMode.inactive
+    @State internal var editMode = EditMode.inactive
     @State private var showingEndSessionAlert = false
     
     var body: some View {
@@ -21,20 +21,12 @@ struct MobSessionView: View {
                     if vm.isEditing {
                         ConfigureSessionView()
                     } else {
-                        RotationLabel()
-                        if vm.mobTimer.isTimerRunning {
-                            TimerCountdown()
-                        } else {
-                            TimerView()
-                        }
+                        MobTimerView(isTimerRunning: vm.mobTimer.isTimerRunning)
                     }
                     HStack {
                         infoButton
                         Spacer()
-                        
-                        if !vm.mobTimer.isTimerRunning {
-                            shuffleButton
-                        }
+                        shuffleButton
                     }
                 }
                 
@@ -74,55 +66,6 @@ struct MobSessionView: View {
                 vm.isKeyboardPresented = value
             }
             .animation(.spring(), value: vm.isKeyboardPresented)
-        }
-    }
-    
-    var infoButton: some View {
-        SymbolButton(action: {
-            showingInfoSheet = true
-        }, symbolName: "info.circle", color: .mobOrange)
-        .font(.title)
-        .padding(.leading, 30)
-    }
-    
-    var shuffleButton: some View {
-        SymbolButton(action: {
-            withAnimation {
-                vm.shuffleTeam()
-            }
-        }, symbolName: "shuffle", color: .mobOrange)
-        .font(.title)
-        .padding(.trailing, 30)
-    }
-}
-
-// MARK: Navigation Toolbar Items
-extension MobSessionView {
-    var logo: some ToolbarContent {
-        ToolbarItem(placement: .principal) {
-            Image("MobProLogo")
-        }
-    }
-    
-    var toggleSettingsButton: some ToolbarContent {
-        ToolbarItem(placement: .navigationBarTrailing) {
-            Button(action: {
-                withAnimation {
-                    if editMode == .inactive {
-                        editMode = .active
-                    } else if editMode == .active{
-                        editMode = .inactive
-                    }
-                    vm.isEditing.toggle()
-                }
-            }, label: {
-                if vm.isEditing {
-                    Text("Save")
-                        .foregroundColor(.mobGreen)
-                } else {
-                    Image(systemName: "gear")
-                }
-            })
         }
     }
 }
